app[REFACT](VM) moved stripe fetchs in VM
